==> Mouse/01_Mouse_exonic_miR_curation_ 122224.ipynb <==
### Mouse data ( 1st step of data cleaning: reshape column)

import pandas as pd
import os
import re
import matplotlib.pyplot as plt

current_dir=os.getcwd()

import import_ipynb
from utils.Data import MOUSE

mouse=MOUSE(f'{current_dir}/UCSC_data_Mouse/mouse_genes_020324_NCBI.csv')

mouse.col_preprocess()

mouse.drop()

mouse.df.rename(columns={'name2':'gene_name'},inplace=True)
### Replace the column name

# Function to calculate mRNA length
def calculate_mrna_length(row):
    starts = list(map(int, row['exonStarts'].split(',')[:-1]))
    ends = list(map(int, row['exonEnds'].split(',')[:-1]))
    lengths = [end - start for start, end in zip(starts, ends)]
    return sum(lengths)

# Apply the function to each row and create a new column for mRNA length
mouse.df['mature_mRNA_Length'] = mouse.df.apply(calculate_mrna_length, axis=1)

mouse_gene_df=mouse.select(['chrom','exonStarts','exonCount',
                        'gene_name','name'])

mouse_gene_df['exonStarts']=mouse_gene_df['exonStarts'].str.replace(r',\Z','',regex=True)
### remove the comma by the end of the column prior to splitting the data

mouse_gene_df['exonStarts']=mouse_gene_df['exonStarts'].str.split(",")

mouse_gene_df=mouse_gene_df.explode('exonStarts')

mouse_gene_df=mouse_gene_df.reset_index(drop=True)

mouse_gene_df_2=mouse.select(['exonEnds','cdsStart','cdsEnd','strand','exonCount','txStart','txEnd','mature_mRNA_Length','gene_name','name'])

mouse_gene_df_2['exonEnds']=mouse_gene_df_2['exonEnds'].str.replace(r',\Z','',regex=True)
### remove the comma at the end of column

mouse_gene_df_2['exonEnds']=mouse_gene_df_2['exonEnds'].str.split(",")

mouse_gene_df_2=mouse_gene_df_2.explode('exonEnds')

mouse_gene_df_2.head(50).reset_index(drop=True)

mouse_gene_df_2.rename(columns={'gene_name':'Name','exonCount':'exon_count','name':'ID'},inplace=True)

mouse_gene_df_2 = mouse_gene_df_2.reset_index(drop=True)

### Concatanate the dataframe (Reshape the dataframe) and remove all miRNA dataset

finalized_mouse_gene_df=pd.concat([mouse_gene_df,mouse_gene_df_2], axis=1)#### This file contains all mouse genes!

finalized_mouse_gene_df.shape

finalized_mouse_gene_df.columns

finalized_mouse_gene_df.drop(columns=['exonCount','gene_name','name'],inplace=True)

finalized_mouse_gene_df.columns

####This regex only extract microRNA but not lncRNA.
mouse_p=r'^Mir\d+(-\d+)*$|^Mir\d+[a-z](-\d+)*$|^Mir\d+[a-z]([a-z]|\d)*$(?<!hg)$|^Mirlet\d[a-z]-*\d*$(?<!hg)$'

mouse_all_genes_no_miR_df=finalized_mouse_gene_df.loc[~finalized_mouse_gene_df['Name'].str.contains(
mouse_p, flags=re.I,regex=True)]####Remove all miR from the data. So this file has all mouse genes except for miRNAs

mouse_all_genes_no_miR_df.shape

mouse_all_genes_no_miR_df.columns

### save the file
mouse_all_genes_no_miR_df.to_csv(f'{current_dir}/Mouse_bash_script/mouse_all_genes_no_miR_df_NCBI.tsv',sep='\t',index=False)

### Extract miRNA data

##extract mouse miRs
mouse_df_miR=finalized_mouse_gene_df.loc[finalized_mouse_gene_df['Name'].str.contains(mouse_p,flags=re.I,regex=True)]

mouse_df_miR.shape ### this dataset do not have lncRNA (e.g. miR-210HG) It has only miRNAs

mouse_df_miR.head(5)

###save the file for microRNA dataset
mouse_df_miR.to_csv(f'{current_dir}/Mouse_bash_script/df_mousemiR_NCBI.tsv',sep="\t",index=False)

## 6. Before you run bedtools_intronic_miRs.sh, reshape the dataframe first and map intronic miRNA with the following steps

The reason for reshaping the dataframe are the following:
* We need to use txStart, txEnd coordinates instead of exon coordinates to find intronic miRNAs. 
* The finalized tsv files will be "mouse_all_genes_no_miR_df_TX_loc_NCBI.tsv".
* We will use this new tsv file to overlap the coordinates of miR to map all the intronic miRNAs.
* Remember the new file has txStart and txEnd locations instead of exon locations.
* bedtools_intronic_miRs.sh will also map the rest of miRNAs that don't have mRNA host genes.
