==> Mouse/02_Mouse_miR_preprocessing_122024.ipynb <==
## Step#2
### This step is to organize your exonic microRNA files
* we only want to have the primary miRNA name and their host genes
(Run this section of code after running bedtools_mouse_exonic_miR.sh)

## Organize the output file(exonic miRs)
* Asssign gene types (noncoding vs protein-coding genes).
* Map the regions of exonic micrornas (5'UTR, 3'UTR, CDS).
* Map all the locations of miRs (exonic, intronic, no host mRNA, span intron-exon.)

### Mouse

#bedtools will generate the overlapped exonic miR files--->mouse_exonic_miR_NCBI.tsv
import pandas as pd
import os
import re
columns=['hostgene_chrom', 'hostgene_exonStarts', 'hostgene_exonEnds', 'hostgene_cdsStart', 'hostgene_cdsEnd'
         , 'hostgene_strand','hostgene_exonCounts','hostgene_txStarts','hostgene_txEnds','mature_host_mRNA_Length',
       'hostgene_Name', 'hostgene_NCBI_ID','miR_chrom', 'miR_exonStarts', 'miR_exonEnds', 'miR_type'
         , 'miR_Name',
       'miR_strand', 'miR_ID']

current_dir=os.getcwd() # Get current directory

# Organize and finalize the exonic microRNA files from the bedtools result

mouse_exonicmiR_df=pd.read_csv(f'{current_dir}/Mouse_tsv_file/mouse_exonic_miR_NCBI.tsv', sep='\t', header=None, names=columns)

mouse_exonicmiR_df.head()

### Extract only primary miR transcript
mouse_exonicmiR_df=mouse_exonicmiR_df.loc[mouse_exonicmiR_df['miR_type']=="miRNA_primary_transcript"]

mouse_exonicmiR_df.shape

#Change string values in the columns of exonic miR files

# Reset the index
mouse_exonicmiR_df = mouse_exonicmiR_df.reset_index(drop=True)

# Create a new DataFrame with the updated "miR_name" column by remove "Name=" from each box
new_df = pd.DataFrame({"miR_name": mouse_exonicmiR_df["miR_Name"].str.replace("Name=", "")})

# Concatenate the new DataFrame with the remaining columns from the original DataFrame
mouse_exonicmiR_df.insert(12,"miR_name", new_df['miR_name'])
mouse_exonicmiR_df.drop(columns=["miR_Name"],inplace=True)

mouse_exonicmiR_df.head()

mouse_exonicmiR_df.shape

csv_file_path=f'{current_dir}/Mouse_csv_file'

#save the mouse exonic data (tsv) file as a csv file
mouse_exonicmiR_df.to_csv(f'{csv_file_path}/mouse_exonic_miR_NCBI.csv',index=False)
#convert to csv file

## Before you run bedtools_intronic_miRs.sh, reshape the dataframe first and map intronic miRNA with the following steps

The reason for reshaping the dataframe are the following:
* We need to use txStart, txEnd coordinates instead of exon coordinates to find intronic miRNAs. 
* The finalized tsv files will be "mouse_all_genes_no_miR_df_TX_loc_NCBI.tsv".
* We will use this new tsv file to overlap the coordinates of miR to map all the intronic miRNAs.
* Remember the new file has txStart and txEnd locations instead of exon locations.
* bedtools_intronic_miRs.sh will also map the rest of miRNAs that don't have mRNA host genes.

import import_ipynb
from utils.Data import MOUSE

### First do the same steps that you did for exonic miRNAs to reshape the dataframe
* ##### You don't have to use exon coordinates of the host genes 

mouse=MOUSE(f'{current_dir}/UCSC_data_Mouse/mouse_genes_020324_NCBI.csv')#mouse gene coordinates tables

mouse.df.columns

"""This function reorganize and clean the columns by doing the following:"""
mouse.col_preprocess()
# mouse.df.columns=gene.df.columns.str.replace("#","") 
### Get rid of "#" symbols in the column names.
# gene.df.columns=gene.df.columns.str.replace('"',"")
### Remove '' in the column names.

mouse.drop()### drop 'bin' column

mouse.df.rename(columns={'name2':'gene_name'},inplace=True)
### Replace the column name

# Function to calculate mature mRNA length
def calculate_mrna_length(row):
    starts = list(map(int, row['exonStarts'].split(',')[:-1]))
    ends = list(map(int, row['exonEnds'].split(',')[:-1]))
    lengths = [end - start for start, end in zip(starts, ends)]
    return sum(lengths)

# Apply the function to each row and create a new column for mRNA length
mouse.df['mature_mRNA_Length'] = mouse.df.apply(calculate_mrna_length, axis=1)

#### This dataframe has all the transctiption starts and ends coordinates
mouse_gene_df_tx=mouse.select(['chrom','txStart','txEnd','cdsStart','cdsEnd','strand',
                        'gene_name','name'])

mouse_gene_df_tx.rename(columns={'gene_name':'Name','name':'ID'},inplace=True)

#Use the same regex again
mouse_p=r'^Mir\d+(-\d+)*$|^Mir\d+[a-z](-\d+)*$|^Mir\d+[a-z]([a-z]|\d)*$(?<!hg)$|^Mirlet\d[a-z]-*\d*$(?<!hg)$'

#### Unlike the previous file, this file has txStart and txEnd coordinates
mouse_all_genes_no_miR_df_TX_loc=mouse_gene_df_tx.loc[~mouse_gene_df_tx['Name'].str.contains(
mouse_p, flags=re.I,regex=True)]####Remove all miR from the data.

### this file has all the transcription starts and ends coordinates but no miRNAs
mouse_all_genes_no_miR_df_TX_loc.to_csv(f'{current_dir}/Mouse_tsv_file/mouse_all_genes_no_miR_df_TX_loc_NCBI.tsv',sep='\t',index=False)

#### Run bedtools_mouse_nonexonic_miR.sh and then bedtools_mouse_intronic_miR.sh

##### use this file, bedtools_mouse_intronic_miR.sh, to find intronic miRNAs.
