==> Mouse/03_Mouse_intronic_miR_organization.ipynb <==
## Intronic miRNAs organization

#bedtools will generate the overlapped intronic miR files. The file to use is bedtools_mouse_intronic_miR.sh
import pandas as pd
import os
current_dir=os.getcwd()
columns_2=['hostgene_chrom', 'hostgene_TxStarts', 'hostgene_TxEnds', 'hostgene_cdsStart', 'hostgene_cdsEnd'
         , 'hostgene_strand',
       'hostgene_Name', 'hostgene_NCBI_ID','miR_chrom', 'miR_exonStarts', 'miR_exonEnds', 'miR_type'
         , 'miR_Name',
       'miR_strand', 'miR_ID']

# Organize and finalize the intronic microRNA files from the bedtools result

mouse_intronicmiR_df=pd.read_csv(f'{current_dir}/Mouse_tsv_file/mouse_intronic_miR_NCBI.tsv', sep='\t', header=None, names=columns_2)

mouse_intronicmiR_df.head()

mouse_intronicmiR_df.shape

### Extract only primary miR transcript
mouse_intronicmiR_df=mouse_intronicmiR_df.loc[mouse_intronicmiR_df['miR_type']=="miRNA_primary_transcript"]

mouse_intronicmiR_df.shape

#Change string values in the columns of exonic miR files

# Reset the index
mouse_intronicmiR_df= mouse_intronicmiR_df.reset_index(drop=True)

# Create a new DataFrame with the updated "miR_name" column
mm_intron_new_df = pd.DataFrame({"miR_name": mouse_intronicmiR_df["miR_Name"].str.replace("Name=", "")})

# Concatenate the new DataFrame with the remaining columns from the original DataFrame
mouse_intronicmiR_df.insert(12,"miR_name", mm_intron_new_df ['miR_name'])
mouse_intronicmiR_df.drop(columns=["miR_Name"],inplace=True)

mouse_intronicmiR_df.head()

mouse_intronicmiR_df.shape

csv_file_path=current_dir+"/Mouse_csv_file"

#save it as a csv file
mouse_intronicmiR_df.to_csv(f'{csv_file_path}/mouse_intronic_miR_NCBI.csv',index=False)#convert it to csv file
