==> Mouse/04_Mouse_miR_host_gene_type_miR.ipynb <==
## Use Gene_func file to add additional columns to the dataframe 
* This new column will have a new column called gene type. Bascially this new dataframe will tell you if the host gene is classified as a protein or non-coding gene. Also, this is applied to both exonic and intronic miRNAs.

### Exonic miRNAs

import pandas as pd

import import_ipynb

import os
current_dir=os.getcwd()

csv_file_path=f'{current_dir}/Mouse_csv_file'

from utils.Gene_func import Gene_type, Cal_mRNA_length

### This dataframe has all exonic miR and hose genes (isoforms); I haven't dropped duplicates yet.
mouse_exonicmiR_df=pd.read_csv(f'{csv_file_path}/mouse_exonic_miR_NCBI.csv')

mouse_exonicmiR_df

mouse_exonicmiR_df.shape

Gene_type(mouse_exonicmiR_df,'hostgene_cdsStart','hostgene_cdsEnd','gene_type')

Cal_mRNA_length(mouse_exonicmiR_df, 'hostgene_txStarts','hostgene_txEnds','hostgene_cdsStart','hostgene_cdsEnd','gene_type','pre_mRNA_length')

mouse_exonicmiR_df.head()

def move_column(df, column, new_position):
    """This function will reassign the column to a new position."""
    col = df.pop(column)
    df.insert(new_position, column, col)
    return df

move_column(mouse_exonicmiR_df,'pre_mRNA_length',9)

### save it as a csv file
mouse_exonicmiR_df.to_csv(f'{csv_file_path}/mouse_exonic_miR(gene_type)_NCBI.csv',index=False)

### Intronic miRNAs

### This dataframe has all intronic miR and hose genes (isoforms); I haven't dropped any duplicates yet.
mouse_intronicmiR_df=pd.read_csv(f'{csv_file_path}/mouse_intronic_miR_NCBI.csv')

Gene_type(mouse_intronicmiR_df,'hostgene_cdsStart','hostgene_cdsEnd','gene_type')

mouse_intronicmiR_df.head()

mouse_intronicmiR_df.shape

"""Calculate the premRNA length of the host genes for each embedded intronic miRNAs"""
Cal_mRNA_length(mouse_intronicmiR_df,'hostgene_TxStarts','hostgene_TxEnds',
                'hostgene_cdsStart','hostgene_cdsEnd','gene_type','pre_mRNA_length' )

mouse_intronicmiR_df.head()

move_column(mouse_intronicmiR_df,'pre_mRNA_length',5)

mouse_intronicmiR_df.to_csv(f'{csv_file_path}/mouse_intronic_miR(gene_type)_NCBI.csv',index=False)
